# packages/mtmai/mtmai-0.3.1218-py3-none-any.whl/mtmai/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "mtmai",
    "description": "mtmai description(group)",
    "version": "0.3.1218"
  },
  "paths": {
...
    "/api/v1/agent/agent_bootstrap": {
      "get": {
        "tags": [
          "agent"
        ],
        "summary": "Agent Bootstrap",
        "description": "\u83b7\u53d6 agent \u7684\u914d\u7f6e\uff0c\u7528\u4e8e\u524d\u7aef\u52a0\u8f7dagent\u7684\u914d\u7f6e",
        "operationId": "agent-agent_bootstrap",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentBootstrap"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/agent/image/{agent}": {
      "get": {
        "tags": [
          "agent"
        ],
        "summary": "\u83b7\u53d6\u5de5\u4f5c\u6d41\u56fe\u50cf",
        "description": "\u6b64\u7aef\u70b9\u901a\u8fc7\u7ed9\u5b9a\u7684 agent ID\uff0c\u751f\u6210\u5de5\u4f5c\u6d41\u7684\u56fe\u50cf\u5e76\u8fd4\u56de PNG \u683c\u5f0f\u7684\u6570\u636e\u3002",
        "operationId": "agent-ChatMessagesItem",
        "parameters": [
          {
            "name": "role",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 64
                },
                {
                  "type": "null"
                }
              ],
              "title": "Role"
            }
          },
          {
            "name": "content",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Content"
            }
          },
          {
            "name": "component",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 64
                },
                {
                  "type": "null"
                }
              ],
              "title": "Component"
            }
          },
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Body_agent-ChatMessagesItem"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "\u8fd4\u56de PNG \u683c\u5f0f\u7684\u5de5\u4f5c\u6d41\u56fe\u50cf\u3002",
            "content": {
              "application/json": {
                "schema": {}
              },
              "image/png": {}
            }
          },
          "404": {
            "description": "Agent \u672a\u627e\u5230"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agent": {
      "get": {
        "tags": [
          "agent"
        ],
        "summary": "\u83b7\u53d6 Agent \u5217\u8868",
        "description": "\u6b64\u7aef\u70b9\u7528\u4e8e\u83b7\u53d6 agent \u5217\u8868\u3002\u652f\u6301\u5206\u9875\u529f\u80fd\u53ef\u4ee5\u901a\u8fc7 `skip` \u548c `limit` \u53c2\u6570\u63a7\u5236\u8fd4\u56de\u7684 agent \u6570\u91cf\u3002",
        "operationId": "agent-ChatMessagesItem",
        "parameters": [
          {
            "name": "role",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 64
                },
                {
                  "type": "null"
                }
              ],
              "title": "Role"
            }
          },
          {
            "name": "content",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Content"
            }
          },
          {
            "name": "component",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 64
                },
                {
                  "type": "null"
                }
              ],
              "title": "Component"
            }
          },
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Body_agent-ChatMessagesItem"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "\u6210\u529f\u8fd4\u56de agent \u5217\u8868",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentsPublic"
                },
                "example": {
                  "data": [
                    {
                      "name": "agent1",
                      "status": "active"
                    },
                    {
                      "name": "agent2",
                      "status": "inactive"
                    }
                  ],
                  "count": 2
                }
              }
            }
          },
          "401": {
            "description": "\u672a\u7ecf\u6388\u6743\u7684\u8bf7\u6c42"
          },
          "500": {
            "description": "\u670d\u52a1\u5668\u5185\u90e8\u9519\u8bef"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agent/chat_messages": {
      "post": {
        "tags": [
          "agent"
        ],
        "summary": "Messages",
        "description": "\u83b7\u53d6\u804a\u5929\u6d88\u606f",
        "operationId": "agent-messages",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentChatMessageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatMessagesResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/form/schema_form/{name}": {
      "get": {
        "tags": [
          "form"
        ],
        "summary": "Schema Form",
        "operationId": "form-schema_form",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MtForm"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/form/submit_schema_form": {
      "post": {
        "tags": [
          "form"
        ],
        "summary": "Submit Schema Form",
        "description": "\u63a5\u53d7\u52a8\u6001\u8868\u5355\u63d0\u4ea4\uff0c\u89e6\u53d1\u5de5\u4f5c\u6d41\u8fd0\u884c\uff0c\u5e76\u6d41\u5f0f\u56de\u4f20 http stream \u4e8b\u4ef6",
        "operationId": "form-submit_schema_form",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AgentBootstrap": {
        "properties": {
          "view_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AgentViewType"
              },
              {
                "type": "null"
              }
            ],
            "default": "sidebar"
          },
          "is_show_fab": {
            "type": "boolean",
            "title": "Is Show Fab",
            "default": true
          }
        },
        "type": "object",
        "title": "AgentBootstrap",
        "description": "\u524d\u7aef\u83b7\u53d6 agent \u7684\u914d\u7f6e\n\u524d\u7aef\u5168\u5c40agent \u52a0\u8f7d\u5668\u4f1a\u5728\u6240\u6709\u9875\u9762\u52a0\u8f7d\u65f6\u8fd4\u56de\u914d\u7f6e\u521d\u59cb\u5316agent UI"
      },
      "AgentChatMessageRequest": {
        "properties": {
          "chat_id": {
            "type": "string",
            "title": "Chat Id"
          },
          "skip": {
            "type": "integer",
            "title": "Skip",
            "default": 0
          },
          "limit": {
            "type": "integer",
            "title": "Limit",
            "default": 100
          }
        },
        "type": "object",
        "required": [
          "chat_id"
        ],
        "title": "AgentChatMessageRequest"
      },
      "AgentMeta": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "base_url": {
            "type": "string",
            "title": "Base Url"
          },
          "chat_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chat Url"
          },
          "can_chat": {
            "type": "boolean",
            "title": "Can Chat",
            "default": [
              false
            ]
          },
          "agent_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent Type"
          },
          "graph_image": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Graph Image"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "base_url"
        ],
        "title": "AgentMeta"
      },
      "AgentViewType": {
        "type": "string",
        "enum": [
          "sidebar",
          "popup"
        ],
        "title": "AgentViewType"
      },
      "AgentsPublic": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/AgentMeta"
            },
            "type": "array",
            "title": "Data"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "data",
          "count"
        ],
        "title": "AgentsPublic"
      },
...
      "Body_agent-ChatMessagesItem": {
        "properties": {
          "props": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Props"
          },
          "artifacts": {
            "anyOf": [
              {
                "items": {
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artifacts"
          }
        },
        "type": "object",
        "title": "Body_agent-ChatMessagesItem"
      },
...
      "ChatMessagesItem": {
        "properties": {
          "role": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Role"
          },
          "content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content"
          },
          "component": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Component"
          },
          "props": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Props"
          },
          "artifacts": {
            "anyOf": [
              {
                "items": {
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artifacts"
          },
          "id": {
            "type": "string",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "id"
        ],
        "title": "ChatMessagesItem"
      },
      "ChatMessagesResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/ChatMessagesItem"
            },
            "type": "array",
            "title": "Data"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "data",
          "count"
        ],
        "title": "ChatMessagesResponse"
      },
...
      "FormFieldSchema": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "placeholder": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Placeholder"
          },
          "valueType": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valuetype"
          },
          "defaultValue": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Defaultvalue"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Type"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "FormFieldSchema"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MtForm": {
        "properties": {
          "properties": {
            "additionalProperties": {
              "$ref": "#/components/schemas/FormFieldSchema"
            },
            "type": "object",
            "title": "Properties"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "type": {
            "type": "string",
            "title": "Type",
            "default": "object"
          },
          "variant": {
            "type": "string",
            "enum": [
              "default"
            ],
            "const": "default",
            "title": "Variant",
            "default": "default"
          }
        },
        "type": "object",
        "required": [
          "properties",
          "title"
        ],
        "title": "MtForm"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/login/access-token"
          }
        }
      }
    }
  },
...
}